requirements.txt
Django==5.0.1
djangorestframework==3.14.0
celery==5.3.6
feedparser==6.0.10
python-dateutil==2.8.2
openai==1.12.0
spacy==3.7.2
python-dotenv==1.0.0
requests==2.31.0
django-filter==23.5
psycopg2-binary==2.9.9
redis==5.0.1
